+++ example/lib/screens/forms/text_box.dart
@@ -29,7 +29,12 @@ class TextBoxPage extends ScrollablePage {
         child: Row(children: const [
           Expanded(child: TextBox()),
           SizedBox(width: 10.0),
-          Expanded(child: TextBox(enabled: false, placeholder: 'Disabled'))
+          Expanded(
+            child: TextBox(
+              enabled: false,
+              placeholder: 'Disabled TextBox',
+            ),
+          )
         ]),
         codeSnippet: '''TextBox()''',
       ),
@@ -44,11 +49,13 @@ class TextBoxPage extends ScrollablePage {
             expands: false,
           ),
         ),
-        codeSnippet: '''TextBox(
-  header: 'Enter your name:',
-  placeholder: 'Name',
-  expands: false,
-),''',
+        codeSnippet: '''InfoLabel(
+  label: 'Enter your name:',
+  child: const TextBox(
+    placeholder: 'Name',
+    expands: false,
+  ),
+)''',
       ),
       subtitle(
         content: const Text('A read-only TextBox with various properties set'),
@@ -60,7 +67,7 @@ class TextBoxPage extends ScrollablePage {
           style: TextStyle(
             fontFamily: 'Arial',
             fontSize: 24.0,
-            letterSpacing: 8,
+            letterSpacing: 8.0,
             color: Color(0xFF5178BE),
             fontStyle: FontStyle.italic,
           ),
@@ -71,7 +78,7 @@ class TextBoxPage extends ScrollablePage {
   style: TextStyle(
     fontFamily: 'Arial,
     fontSize: 24.0,
-    letterSpacing: 8,
+    letterSpacing: 8.0,
     color: Color(0xFF5178BE),
     fontStyle: FontStyle.italic,
   ),