+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add `TextFormBox.initialValue` ([#749](https://github.com/bdlukaa/fluent_ui/issues/749))
 - Add `PaneItem.enabled` ([#748](https://github.com/bdlukaa/fluent_ui/discussions/748))
+- Add Thai localization
 
 ## 4.4.0